// apps/web/src/lib/api/generated/models/screeningQueueItemDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export interface ScreeningQueueItemDto {
	/** @nullable */
	abstract_text?: string | null;
	/** @nullable */
	doi?: string | null;
	final_status: string;
	full_text_status: string;
	/** @nullable */
	publication_year?: number | null;
	report_id: string;
	revision: number;
	/** @nullable */
	title?: string | null;
	title_abstract_status: string;
}